+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeClient.kt
@@ -137,9 +137,9 @@ open class BraintreeClient @VisibleForTesting internal constructor(
         configuration: Configuration?,
         authorization: Authorization
     ) {
-        if (isAnalyticsEnabled(configuration)) {
+        configuration?.let {
             analyticsClient.sendEvent(
-                configuration!!,
+                it,
                 eventName,
                 sessionId,
                 integrationType,
@@ -323,7 +323,15 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      */
     @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     fun reportCrash() =
-        analyticsClient.reportCrash(applicationContext, sessionId, integrationType, authorization)
+        getConfiguration { configuration, _ ->
+            analyticsClient.reportCrash(
+                applicationContext,
+                configuration,
+                sessionId,
+                integrationType,
+                authorization
+            )
+        }
 
     // NEXT MAJOR VERSION: Make launches browser switch as new task a property of `BraintreeOptions`
     fun launchesBrowserSwitchAsNewTask(): Boolean {
@@ -345,16 +353,5 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      */
     open fun launchesBrowserSwitchAsNewTask(launchesBrowserSwitchAsNewTask: Boolean) {
         this.launchesBrowserSwitchAsNewTask = launchesBrowserSwitchAsNewTask
-    }
-
-    companion object {
-
-        /**
-         * @suppress
-         */
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-        fun isAnalyticsEnabled(configuration: Configuration?): Boolean {
-            return configuration != null && configuration.isAnalyticsEnabled
-        }
     }
 }