+++ CHANGELOG.md
@@ -2,15 +2,18 @@
 
 ## unreleased
 
+* Move from Braintree to PayPal analytics service (FPTI)
 * Breaking Changes
   * All Modules
     * Bump `minSdkVersion` to API 23
     * Bump target Java version to Java 11
-    * Update payment method constructor parameters from `braintreeClient` to `context` and
-      `authorization`
   * BraintreeCore
     * Remove `BraintreeClient` public constructors
     * Remove `ClientTokenProvider`
+    * Remove `BraintreeClient` public constructors
+    * Update payment method constructor parameters from `braintreeClient` to `context` and 
+      `authorization`
+    * Remove `BraintreeSharedPreferencesException`
   * UnionPay
     * Remove `union-pay` module
       * UnionPay cards can now be processed as regular cards (through the `card` module) due to their partnership with Discover
@@ -39,6 +42,7 @@
     * Change `GooglePayClient#requestPayment` parameters and rename to 
       `GooglePayClient#createPaymentAuthRequest`
     * Add `GooglePayClient#tokenize` 
+    * Remove `merchantId` from `GooglePayRequest`
   * ThreeDSecure
     * Remove `ThreeDSecureListener`
     * Add `ThreeDSecureLauncher`, `ThreeDSecurePaymentAuthResult`, and `ThreeDSecureLancherCallback`
@@ -52,6 +56,7 @@
     * Rename `ThreeDSecureResult#tokenizedCard` to `ThreeDSecureResult#threeDSecureNonce`
     * Remove `ThreeDSecureV1UICustomization`
     * Remove `versionRequesed` from `ThreeDSecureRequest` as version 1 is no longer supported
+    * Remove `ThreeDSecureV2BaseCustomization`
   * PayPal
     * Remove `PayPalListener`
     * Add `PayPalLauncher`, `PayPalPaymentAuthRequest`, `PayPalPaymentAuthResult`,